Adding interpolation search and the Big O cases

+++ search_algorithms/102-interpolation.c
@@ -11,7 +11,7 @@
  */
 int interpolation_search(int *array, size_t size, int value)
 {
-	unsigned int low = 0, pos = 0, high = size - 1;
+	size_t low = 0, pos = 0, high = size - 1;
 
 	if (!array)
 		return (-1);
